Factor compatibility_distance into its own module
Computing of compatibility distance is a separate concern from the
genome itself. Now I can document and think about it in a
dedicated place.
